feat : 기본 세팅
