+++ Common/src/main/java/com/seveneleven/util/security/service/ProjectAuthorizationCheckService.java
@@ -1,0 +1,4 @@
+package com.seveneleven.util.security.service;
+
+public interface ProjectAuthorizationCheckService {
+}
