+++ Common/src/main/java/com/seveneleven/util/security/service/ProjectAuthorizationCheckServiceImpl.java
@@ -1,0 +1,9 @@
+package com.seveneleven.util.security.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class ProjectAuthorizationCheckServiceImpl {
+
+    // TODO - 프로젝트 권한 체크 서비스 - For @PreAuthorization - SpEL 사용
+}
